- fixed a bug in FCDoubleSpinner GUI element - added a new parameter in NCC tool named offset. If the offset is used then the copper clearing will finish to a set distance of the copper features

+++ flatcamGUI/GUIElements.py
@@ -1609,7 +1609,7 @@ class FCDoubleSpinner(QtWidgets.QDoubleSpinBox):
 
     def set_value(self, val):
         try:
-            k = int(val)
+            k = float(val)
         except Exception as e:
             log.debug(str(e))
             return
